+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 1.2.2 (2017-03-02)
+
+- Improved browser support:
+    - Object.assign replacement
+    - Removed forEach calls on NodeList objects
+- Started works on a new demo
+
 ## 1.2.1 (2017-03-01)
 
 - README update